requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
pandas
openpyxl
celery
redis
python-dotenv
python-dateutil